take account of UI element when wrapping head with wrapRootElement in SSR
