+++ packages/gatsby/cache-dir/head/head-export-handler-for-ssr.js
@@ -4,12 +4,11 @@ const { StaticQueryContext } = require(`gatsby`)
 const {
   headExportValidator,
   filterHeadProps,
-  warnForInvalidTags,
+  getValidHeadNodeForSSR,
 } = require(`./utils`)
 const { ServerLocation, Router } = require(`@gatsbyjs/reach-router`)
 const { renderToString } = require(`react-dom/server`)
 const { parse } = require(`node-html-parser`)
-const { VALID_NODE_NAMES } = require(`./constants`)
 import { apiRunner } from "../api-runner-browser"
 
 export function headHandlerForSSR({
@@ -36,7 +35,7 @@ export function headHandlerForSSR({
         <pageComponent.Head {...filterHeadProps(_props)} />
       )
 
-      const headWithWrapRootElment = apiRunner(
+      const headWithWrapRootElement = apiRunner(
         `wrapRootElement`,
         { element: <HeadElement /> },
         <HeadElement />,
@@ -45,7 +44,7 @@ export function headHandlerForSSR({
         }
       ).pop()
 
-      return headWithWrapRootElment
+      return headWithWrapRootElement
     }
 
     const routerElement = (
@@ -61,55 +60,9 @@ export function headHandlerForSSR({
       </StaticQueryContext.Provider>
     )
 
-    // extract head nodes from string
     const rawString = renderToString(routerElement)
-    const headNodes = parse(rawString).childNodes
-
-    const validHeadNodes = []
-
-    const seenIds = new Map()
-    for (const node of headNodes) {
-      const { rawTagName } = node
-      const id = node.attributes?.id
-
-      if (!VALID_NODE_NAMES.includes(rawTagName)) {
-        warnForInvalidTags(rawTagName)
-      } else {
-        let element
-        const attributes = { ...node.attributes, "data-gatsby-head": true }
-        if (rawTagName === `script`) {
-          element = (
-            <script
-              {...attributes}
-              dangerouslySetInnerHTML={{
-                __html: node.text,
-              }}
-            />
-          )
-        } else {
-          element =
-            node.textContent.length > 0 ? (
-              <node.rawTagName {...attributes}>
-                {node.textContent}
-              </node.rawTagName>
-            ) : (
-              <node.rawTagName {...attributes} />
-            )
-        }
-
-        if (id) {
-          if (!seenIds.has(id)) {
-            validHeadNodes.push(element)
-            seenIds.set(id, validHeadNodes.length - 1)
-          } else {
-            const indexOfPreviouslyInsertedNode = seenIds.get(id)
-            validHeadNodes[indexOfPreviouslyInsertedNode] = element
-          }
-        } else {
-          validHeadNodes.push(element)
-        }
-      }
-    }
+    const rootNode = parse(rawString)
+    const validHeadNodes = getValidHeadNodeForSSR(rootNode)
 
     setHeadComponents(validHeadNodes)
   }
